Fix type mismatch in std::max call
- Cast std::thread::hardware_concurrency() to std::size_t to match types
- Fixes compilation error on all platforms where unsigned int != std::size_t
- Ensures both arguments to std::max are the same type (std::size_t)

+++ src/work_stealing_pool.cpp
@@ -339,7 +339,7 @@ std::size_t WorkStealingPool::getOptimalThreadCount() noexcept {
         return logicalCores;
     } else {
         // Should not reach here, but be safe
-        return std::max(std::thread::hardware_concurrency(), std::size_t(2));
+        return std::max(static_cast<std::size_t>(std::thread::hardware_concurrency()), std::size_t(2));
     }
 }
